+++ NEWS.md
@@ -1,5 +1,9 @@
 # testthat 0.X
 
+## Removed functions
+
+* `colourise()` was removed. (Colour is still supported, via the `crayon` package.)
+
 # testthat 0.9
 
 ## New features